--- spateo/external/lack/__init__.py
@@ -1,6 +1,0 @@
-"""__init__ module exports main interfaces that developers of dynamo series packages use. The detailed implementation is based on python logging module with customized wrapper classes."""
-__version__ = "0.0.4"
-
-from .logger import Logger
-from .logger_interface import *
-from .logger_manager import LoggerManager
