--- spateo/external/lack/logger_interface.py
@@ -1,2 +1,0 @@
-from .logger import set_logger_level
-from .logger_manager import LoggerManager
